move files, clean up mutex references
