+++ agent/consul/gateways/controller_routes.go
@@ -1,7 +1,8 @@
-package controller
+package gateways
 
 import (
 	"context"
+	"github.com/hashicorp/consul/agent/consul/controller"
 	"github.com/hashicorp/consul/agent/consul/fsm"
 	"github.com/hashicorp/consul/agent/consul/state"
 	"github.com/hashicorp/consul/agent/consul/stream"
@@ -12,20 +13,18 @@ type tcpRouteReconciler struct {
 	baseReconciler
 }
 
-func (r tcpRouteReconciler) Reconcile(ctx context.Context, req Request) error {
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
+func (r tcpRouteReconciler) Reconcile(ctx context.Context, req controller.Request) error {
 	return nil
 }
 
-func TCPRouteController(fsm *fsm.FSM, publisher state.EventPublisher, logger hclog.Logger) Controller {
+func TCPRouteController(fsm *fsm.FSM, publisher state.EventPublisher, logger hclog.Logger) controller.Controller {
 	reconciler := tcpRouteReconciler{
 		baseReconciler{
 			fsm:    fsm,
 			logger: logger,
 		},
 	}
-	return New(publisher, reconciler).Subscribe(
+	return controller.New(publisher, reconciler).Subscribe(
 		&stream.SubscribeRequest{
 			Topic:   state.EventTopicTCPRoute,
 			Subject: stream.SubjectWildcard,
@@ -37,20 +36,18 @@ type httpRouteReconciler struct {
 	baseReconciler
 }
 
-func (r httpRouteReconciler) Reconcile(ctx context.Context, req Request) error {
-	r.mutex.Lock()
-	defer r.mutex.Unlock()
+func (r httpRouteReconciler) Reconcile(ctx context.Context, req controller.Request) error {
 	return nil
 }
 
-func HTTPRouteController(fsm *fsm.FSM, publisher state.EventPublisher, logger hclog.Logger) Controller {
+func HTTPRouteController(fsm *fsm.FSM, publisher state.EventPublisher, logger hclog.Logger) controller.Controller {
 	reconciler := httpRouteReconciler{
 		baseReconciler{
 			fsm:    fsm,
 			logger: logger,
 		},
 	}
-	return New(publisher, reconciler).Subscribe(
+	return controller.New(publisher, reconciler).Subscribe(
 		&stream.SubscribeRequest{
 			Topic:   state.EventTopicHTTPRoute,
 			Subject: stream.SubjectWildcard,
